+++ backend/src/baserow/config/settings/dev.py
@@ -22,8 +22,7 @@ EMAIL_USE_TLS = False
 EMAIL_HOST = "mailhog"
 EMAIL_PORT = 1025
 
-BASEROW_IMPORT_TOLERATED_TYPE_ERROR_THRESHOLD = 5  # Allow 5% to be more adventurous
-BASEROW_MAX_FILE_IMPORT_ERROR_COUNT = 10  # To trigger this exception easily
+BASEROW_MAX_ROW_REPORT_ERROR_COUNT = 10  # To trigger this exception easily
 
 try:
     from .local import *  # noqa: F403, F401