+++ backend/src/baserow/contrib/database/api/tables/serializers.py
@@ -18,8 +18,15 @@ class TableCreateSerializer(serializers.ModelSerializer):
     data = serializers.ListField(
         min_length=1,
         default=None,
-        help_text="A list of rows that needs to be created as initial table data. If "
-        "not provided some example data is going to be created.",
+        help_text=(
+            "A list of rows that needs to be created as initial table data. "
+            "Each row is a list of values that are going to be added in the new "
+            "table in the same order as provided.\n\n"
+            'Ex: \n```json\n[\n  ["row1_field1_value", "row1_field2_value"],\n'
+            '  ["row2_field1_value", "row2_field2_value"],\n]\n```\n'
+            "for creating a two rows table with two fields.\n\n"
+            "If not provided, some example data is going to be created."
+        ),
     )
     first_row_header = serializers.BooleanField(
         default=False,
@@ -35,6 +42,26 @@ class TableCreateSerializer(serializers.ModelSerializer):
             "data": {"required": False},
             "first_row_header": {"required": False},
         }
+
+
+class TableImportSerializer(serializers.Serializer):
+    data = serializers.ListField(
+        min_length=1,
+        required=True,
+        help_text=(
+            "A list of rows you want to add to the specified table. "
+            "Each row is a list of values, one for each **writable** field. "
+            "The field values must be ordered according to the field order in the "
+            "table. "
+            "All values must be compatible with the corresponding field type.\n\n"
+            'Ex: \n```json\n[\n  ["row1_field1_value", "row1_field2_value"],\n'
+            '  ["row2_field1_value", "row2_field2_value"],\n]\n```\n'
+            "for adding two rows to a table with two writable fields."
+        ),
+    )
+
+    class Meta:
+        fields = ("data",)
 
 
 class TableUpdateSerializer(serializers.ModelSerializer):