+++ backend/src/baserow/contrib/database/table/constants.py
@@ -1,0 +1,1 @@
+TABLE_CREATION = "import-create-table"