+++ backend/tests/baserow/contrib/database/rows/test_rows_actions.py
@@ -1,6 +1,7 @@
 from decimal import Decimal
 
 import pytest
+from unittest.mock import patch
 
 from pytest_unordered import unordered
 
@@ -16,6 +17,7 @@ from baserow.contrib.database.fields.models import SelectOption
 from baserow.contrib.database.rows.actions import (
     CreateRowActionType,
     CreateRowsActionType,
+    ImportRowsActionType,
     DeleteRowActionType,
     DeleteRowsActionType,
     MoveRowActionType,
@@ -235,6 +237,155 @@ def test_can_undo_redo_creating_rows(data_fixture):
     assert_undo_redo_actions_are_valid(action_redone, [CreateRowsActionType])
 
     assert model.objects.all().count() == 3
+
+
+@pytest.mark.undo_redo
+@pytest.mark.django_db
+def test_can_undo_importing_rows(data_fixture):
+    session_id = "session-id"
+    user = data_fixture.create_user(session_id=session_id)
+    table = data_fixture.create_database_table(name="Car", user=user)
+    name_field = data_fixture.create_text_field(
+        table=table, name="Name", text_default="Test", order=1
+    )
+    speed_field = data_fixture.create_number_field(
+        table=table, name="Max speed", number_negative=True, order=2
+    )
+    price_field = data_fixture.create_number_field(
+        table=table,
+        name="Price",
+        number_decimal_places=2,
+        number_negative=False,
+        order=3,
+    )
+    model = table.get_model()
+
+    action_type_registry.get_by_type(ImportRowsActionType).do(
+        user,
+        table,
+        data=[
+            [
+                "Tesla",
+                240,
+                59999.99,
+            ],
+            [
+                "Giulietta",
+                210,
+                34999.99,
+            ],
+            [
+                "Panda",
+                160,
+                8999.99,
+            ],
+        ],
+    )
+
+    assert model.objects.all().count() == 3
+
+    action_undone = ActionHandler.undo(
+        user, [TableActionScopeType.value(table_id=table.id)], session_id
+    )
+
+    assert_undo_redo_actions_are_valid(action_undone, [ImportRowsActionType])
+
+    assert model.objects.all().count() == 0
+
+
+@pytest.mark.django_db
+@pytest.mark.undo_redo
+@patch("baserow.contrib.database.table.signals.table_updated.send")
+@patch("baserow.contrib.database.rows.signals.rows_created.send")
+def test_can_undo_redo_importing_rows(row_send_mock, table_send_mock, data_fixture):
+    session_id = "session-id"
+    user = data_fixture.create_user(session_id=session_id)
+    table = data_fixture.create_database_table(name="Car", user=user)
+    name_field = data_fixture.create_text_field(
+        table=table, name="Name", text_default="Test", order=1
+    )
+    speed_field = data_fixture.create_number_field(
+        table=table, name="Max speed", number_negative=True, order=2
+    )
+    price_field = data_fixture.create_number_field(
+        table=table,
+        name="Price",
+        number_decimal_places=2,
+        number_negative=False,
+        order=3,
+    )
+    model = table.get_model()
+
+    action_type_registry.get_by_type(ImportRowsActionType).do(
+        user,
+        table,
+        data=[
+            [
+                "Tesla",
+                240,
+                59999.99,
+            ],
+            [
+                "Giulietta",
+                210,
+                34999.99,
+            ],
+            [
+                "Panda",
+                160,
+                8999.99,
+            ],
+        ],
+    )
+
+    table_send_mock.assert_called_once()
+    table_send_mock.reset_mock()
+
+    assert model.objects.all().count() == 3
+
+    ActionHandler.undo(
+        user, [TableActionScopeType.value(table_id=table.id)], session_id
+    )
+
+    action_redone = ActionHandler.redo(
+        user, [TableActionScopeType.value(table_id=table.id)], session_id
+    )
+
+    assert_undo_redo_actions_are_valid(action_redone, [ImportRowsActionType])
+
+    assert model.objects.all().count() == 3
+
+    table_send_mock.assert_not_called()
+    row_send_mock.assert_called_once()
+    assert len(row_send_mock.call_args[1]["rows"]) == 3
+
+    # Test that the signal change when we undo more rows
+    action_type_registry.get_by_type(ImportRowsActionType).do(
+        user,
+        table,
+        data=[
+            [
+                "Tesla",
+                240,
+                59999.99,
+            ],
+        ]
+        * 51,
+    )
+
+    row_send_mock.reset_mock()
+    table_send_mock.reset_mock()
+
+    ActionHandler.undo(
+        user, [TableActionScopeType.value(table_id=table.id)], session_id
+    )
+
+    action_redone = ActionHandler.redo(
+        user, [TableActionScopeType.value(table_id=table.id)], session_id
+    )
+
+    table_send_mock.assert_called_once()
+    row_send_mock.assert_not_called()
 
 
 @pytest.mark.django_db